+++ examples/my_code/.recode.exs
@@ -1,6 +1,7 @@
 alias Recode.Task
 
 [
+  version: "0.2.0",
   # Can also be set/reset with "--autocorrect"/"--no-autocorrect".
   autocorrect: true,
   # With "--dry" no changes will be written to the files.